show more info on location sidebar. show warning about scope of changes for location info. Fix delete master file with no image meta. Fix clone of masterfile with location info

+++ app/jobs/delete_master_files.rb
@@ -52,7 +52,7 @@ class DeleteMasterFiles < BaseJob
             end
 
             logger.info "Removing master file record #{del_fn}"
-            mf.image_tech_meta.destroy
+            mf.image_tech_meta.destroy if !mf.image_tech_meta.nil?
             mf.destroy
 
             del_fn = filenames.shift
